kNN Works
All comps get read in to the nearestneighbors method and it predicts on
the new data (small subsection of the data set with just a few binaries
are all that have been tested). Just need to ensure all the right data
is going in and proper values are coming out.